Renaming module to better reflect what it provides.
